==> 03_creditcard/02_creditcard_scripts/04_creditcard_model.Rmd <==
---
title: "Creditcard - Modelling"
author: "Simon Schrauth"
date: "`r Sys.Date()`"
output: html_document
---
# Model Creditcard Dataset

## Goal

After executing this script, a deep neural network for the creditcard dataset should be modeled.

## Preparations

### Load R packages

```{r}
pacman::p_load(tidyverse, 
               here,
               reticulate)
```

### Activate Python environment

```{r}
reticulate::use_condaenv(condaenv = "Masterarbeit", 
                         required = TRUE
                         )
```

### Import Python packages

```{python}
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import tensorflow as tf
from tensorflow import keras
from scikeras.wrappers import KerasClassifier
from sklearn.model_selection import PredefinedSplit
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import confusion_matrix, ConfusionMatrixDisplay
from pathlib import Path
import keras_tuner
```

### Load data
```{r}
cc_X_train_pd = readRDS(file = here("03_creditcard", "01_creditcard_data", "02_creditcard_data_processed",
                                 "03_creditcard_data_model_X_train.rds"))

cc_y_train_pd = readRDS(file = here("03_creditcard", "01_creditcard_data", "02_creditcard_data_processed",
                                 "04_creditcard_data_model_y_train.rds"))

cc_X_test_pd = readRDS(file = here("03_creditcard", "01_creditcard_data", "02_creditcard_data_processed",
                                "05_creditcard_data_model_X_test.rds"))

cc_y_test_pd = readRDS(file = here("03_creditcard", "01_creditcard_data", "02_creditcard_data_processed",
                                "06_creditcard_data_model_y_test.rds"))

cc_col_names_enc = readRDS(file = here("03_creditcard", "01_creditcard_data", "02_creditcard_data_processed",
                                       "07_creditcard_data_model_col_names_enc.rds"))
```

### Transfer dataframes to python
```{python}
cc_X_train = r.cc_X_train_pd
cc_y_train = r.cc_y_train_pd

cc_X_test = r.cc_X_test_pd
cc_y_test = r.cc_y_test_pd

cc_col_names_enc = r.cc_col_names_enc
```

### Convert to Tensor

```{python}
# cc_X_train = tf.convert_to_tensor(cc_X_train_pd)
# cc_y_train = tf.convert_to_tensor(cc_y_train_pd)
# 
# cc_X_val = tf.convert_to_tensor(cc_X_val_pd)
# cc_y_val = tf.convert_to_tensor(cc_y_val_pd)
# 
# cc_X_test = tf.convert_to_tensor(cc_X_test_pd)
# cc_y_test = tf.convert_to_tensor(cc_y_test_pd)
```

## Model Construction

### Preparation
Define certain global variables
```{python}
feature_num = cc_X_train.shape[-1]

batch_size = 32
epochs = 5
```


```{python}


# cc_model = keras.Sequential(
#     [
#         keras.layers.Input(shape = (feature_num,), 
#                            name = "input_layer"),
#         keras.layers.Dense(units = feature_num, 
#                            activation="relu", 
#                            name = "hidden_layer_1"),
#         keras.layers.Dense(units = 10, 
#                            activation="relu", 
#                            name = "hidden_layer_2"),
#         keras.layers.Dense(units = 1, 
#                            activation="sigmoid", 
#                            name = "output_layer"), 
#     ],
#     name = "cc_model"
# )
# 
# cc_model.summary()
```


```{python}
def define_model(units_layer1,
                 units_layer2,
                 learning_rate,
                 momentum
                 ):
                   
  model = keras.Sequential(
    [
        keras.layers.Input(shape = (feature_num,), 
                           name = "input_layer"),

        # keras.layers.Dropout(rate = 0.2, seed = 324095),
                           
        keras.layers.Dense(units = units_layer1,
                           activation = "relu", 
                           name = "hidden_layer_1"),
                           
        # keras.layers.Dropout(rate = 0.3, seed = 957485),
        
        keras.layers.Dense(units = units_layer2, 
                           activation = "relu", 
                           name = "hidden_layer_2"),
                           
        # keras.layers.Dropout(rate = 0.3, seed = 389459),
        
        keras.layers.Dense(units = 1, 
                           activation = "sigmoid", 
                           name = "output_layer"), 
    ]
  )
                      
  model.compile(loss = "binary_crossentropy",
                optimizer = keras.optimizers.SGD(learning_rate = learning_rate,
                                                 momentum = momentum),
                metrics = ["acc"])
                
  return model

def build_model(hp):
  units_layer1 = hp.Int("units_layer1", 
                        min_value = np.around(feature_num*0.5),
                        max_value = np.around(feature_num*2),
                        step = 5)
  
  units_layer2 = hp.Int("units_layer2", 
                        min_value = np.around(feature_num*0.25),
                        max_value = np.around(feature_num*1.25),
                        step = 5)
  
  learning_rate = hp.Float("lr",
                           min_value = 1e-5,
                           max_value = 1e-2,
                           sampling = "log")
                           
  momentum = hp.Float("momentum",
                      min_value = 0.90,
                      max_value = 0.99,
                      step = 0.03)
  
  model = define_model(units_layer1 = units_layer1,
                       units_layer2 = units_layer2,
                       learning_rate = learning_rate,
                       momentum = momentum)
                       
  return model
```

```{python}
build_model(keras_tuner.HyperParameters())
```
```{python}
tuner = keras_tuner.RandomSearch(
  hypermodel = build_model,
  objective = "val_acc",
  max_trials = 10,
  executions_per_trial = 2,
  overwrite = True,
  seed = 59276045,
  project_name = "keras_tuner_model"
)

tuner.search_space_summary()

```

```{python}
tuner.search(cc_X_train,
             cc_y_train,
             epochs = epochs,
             validation_data = (cc_X_test,cc_y_test))
```

```{python}
tuner.results_summary()
best_hp = tuner.get_best_hyperparameters()[0]
best_hp.values

final_model = build_model(best_hp)
final_model.fit(x = cc_X_train, 
                y = cc_y_train)

y_pred = final_model.predict(cc_X_train)

final_model.evaluate(cc_X_test, cc_y_test)

cc_cf = confusion_matrix(y_true = cc_y_train, y_pred = np.rint(final_model.predict(cc_X_train)))

cc_cf_vis = ConfusionMatrixDisplay(cc_cf)
cc_cf_vis.plot(values_format = "")

plt.show()

y_pred_raw = final_model.predict(cc_X_test)
y_true = cc_y_test

y_pred = np.where(y_pred_raw > 0.5, 1, 0)


```


```{python}
# def define_model(activation = "relu", 
#                  units_layer1 = feature_num, 
#                  units_layer2 = np.around(feature_num/2), 
#                  optimizer = "SGD"):
#
# model = KerasClassifier(model = define_model,
#                         activation = "relu",
#                         units_layer1 = feature_num, 
#                         units_layer2 = np.around(feature_num/2), 
#                         loss = "binary_crossentropy",
#                         optimizer = "SGD",
#                         epochs = epochs,
#                         batch_size = batch_size,
#                         verbose = 1)
#                         
# activation = ["relu", "sigmoid"]
# # activation = ["relu"]
# optimizer = ["SGD", "rmsprop", "Adam"]
# # what are those?
# # optimizer = ["SGD"]
# # optimizer__learning_rate = [0.001, 0.01, 0.1] #!!!!!
# 
# units_layer1 = np.around(np.linspace(start = feature_num/2, 
#                                      stop = feature_num/1, 
#                                      num = 3)
#                         )
#    
# units_layer2 = np.around(np.linspace(start = feature_num/4, 
#                                      stop = feature_num/1.25, 
#                                      num = 3)
#                         ) 
# 
# param_grid = dict(activation = activation, 
#                   optimizer = optimizer,
#                   units_layer1 = units_layer1,
#                   units_layer2 = units_layer2)
#                   
# grid = GridSearchCV(estimator = model,
#                     param_grid = param_grid,
#                     scoring = ["accuracy", "roc_auc", "f1", "recall"],
#                     refit = "recall", #!!!!!
#                     cv = cc_ps) 
# 
# grid_result = grid.fit(cc_X_trainval, cc_y_trainval) 
# 
# grid_result.best_score_
# grid_result.best_params_
# grid_result.best_estimator_
# 
# final_model = grid_result.best_estimator_
# 
# final_model.build_fn
# 
# cc_model = final_model.fit(cc_X_trainval, cc_y_trainval, batch_size = batch_size, epochs = epochs)
# 
# cc_results = cc_model.score(cc_X_val, cc_y_val)
# 
# cc_results
# 
# cc_cf = confusion_matrix(y_true = cc_y_train, y_pred = cc_model.predict(cc_X_train))
# 
# cc_cf_vis = ConfusionMatrixDisplay(cc_cf)
# cc_cf_vis.plot()
# 
# plt.show()
```

```{python}
# cc_data_path = Path.cwd().parent.joinpath("01_creditcard_data", "01_creditcard_data_processed")
# 
# cc_model.save(cc_data_path)
```
```{python}
# param = grid_result.best_params_
# 
# xyz = define_model
```
